use dart in `getNextMeetupTime` (#1569)

+++ packages/ew_encointer_utils/lib/src/assigment.dart
@@ -7,8 +7,8 @@ const momentsPerDay = 86400000;
 
 const momentsPerDegree = momentsPerDay / 360;
 
-int meetupTime(double longitude, double attestingStart, double meetupTimeOffset, double oneDay) {
-  final timePerDegree = oneDay / 360;
+int meetupTime(double longitude, int attestingStart, int meetupTimeOffset, int oneDay) {
+  final timePerDegree = oneDay.toDouble() / 360;
   // The meetups start at high sun at 180 degrees and during one day the meetup locations travel
   // along the globe until the very last meetup happens at high sun at -180 degrees.
   // So we scale the range 180...-180 to 0...360.
